requirements.txt
streamlit>=1.24.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
tiktoken>=0.6.0
tenacity>=8.2.3
google-genai
plotly>=5.18.0
reportlab>=4.0.8